refactor(runtime): split engine.rs into per-category module
